Add YamJam to protect sensitive information
Since we are using a public repository on GitHub we need to be careful 
that we aren’t publishing any sensitive information such as database 
credentials or Django’s secret key when checking in code.

YamJam is a python module that allows sensitive information to be stored 
in a YAML file in the project directory. <http://yamjam.readthedocs.org>