Add Django REST Framework (DRF) GIS module
Change the LocationSerializer to use the DRF GIS module for the point
and polygon data.

NOTE: We’re needing to use a git hub trunk version of DRF GIS at the 
moment in order to support DRF v3.1.0. This should be changed to a 
proper release version when available. (see issue #1)