Add production static root to yamjam sample
This should point to the location that static files are to be copied to 
on the production server. Typically this would be somewhere that the 
production web server serves static files from. The collectstatic 
management task should be run once deploying to production to copy the
static assets to this location.